AI-teams-controller-public/backend/experiments/voice_command/pyproject.toml
[project]
name = "voice-command-demo"
version = "0.1.0"
description = "Voice command demos: Speech-to-Text, Text-to-Speech with OpenAI"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "sounddevice>=0.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "langchain>=0.3.0",
    "langchain-xai>=0.2.0",
]

[tool.uv]
dev-dependencies = []
